Update and rename src/me/UseRod/BookGUI/Utils/ConfigVariables.java to src/me/userod/bookgui/utils/ConfigVariables.java
